345.1. Setting Initial Values(Section 24: REST-Based React Apps)
In-class video we read factored stream create to make use of stream form.

We're not going to do the same thing for stream edit as well.

Stream is going to be very similar to the work we just did inside of stream create the only big difference

here remember is that we want to make sure that the form is going to show some initial values inside

of those inputs.

Someone to very quickly tell you how we do that with redux form.

First I need to open up my code editor and find the stream form component.

Here it is right here.

Stream form.

So I just want to remind you very quickly that down at the very bottom of the file when we export a

component from here we have stream form.

And it gets wrapped inside of redux form.

Right.

It gets wrapped by that helper right there.

So just with that in mind I going to show you another diagram here.

Very quickly.

So in this diagram very similar to the one we are just looking at.

I just essentially took off stream creates.

So we have stream edit that is showing stream form wrapped inside of that redux form helper.

So when we passed props from stream at it down to stream form just as we did inside of stream create

remember your stream create right here and we pass some props down to stream form.

We're not actually passing props directly to our stream form component.

Instead we are technically passing props to redux form redux form then turns around and passes those

props onto our component.

So there are actually some very special props that we can pass down into the redux form wrapped component

that we are creating here.

One special component or Simmie one special prop name is initial values.

If we pass some initial values Propp from stream at it down to this kind of redux farm rapt stream form

it's going to provide some initial values to show inside of the text inputs inside of stream form.

So if we pass in a prop named initial values specifically initial values that is a very special name

with redux form.

And if that object contains a title and description then those will be used at the as the initial values

for our form.

Now when I described this with diagrams it probably looks kind of confusing.

So let's just start to work on string at it.

And then when we get to the part where we want to set up these initial values We'll take another glance

at this diagram to make sure you understand what's going on here because this is a really key topic

for understanding how to set up any type of form like this right here with redux form.

So let's get to it.

All right.

So I can open up my code editor and I'm going to find the stream edit component.

Now inside of here we're going to do a step by step refactor just like we did with string create.

So up at the top I've got import connect.

That's good.

I've got import jet stream.

That's good.

I also want to import edit stream from my action Kreator file as well.

Remember edit stream is what is going to eventually allow us to take the idea of a stream and some new

properties for the stream and update our API record or the record maintained by the API.

And then after that I'm also going to import our new stream form component from stream form Zao inside

of your.

We're going to leave the component in MT still want that.

We're going to define an on Samit function.

This is going to be used as our callback for stream form just like we had inside of a stream create.

So when the user eventually submits our form this thing is going to receive some form values and for

right now we will just cancel log out those form values.

Next up inside the render method rather than just returning a div with the stream's title that we fetched

we're going to instead return a multi-line block of GSX just like we did back on stream create.

So now inside if you're going to return a div with an H 3.

Is that what we use on the other one.

When we check really quickly Yeah it's three that says edit a stream and then after that we will place

our stream form component.

Now just like before we're going to pass in an on submit callback and then we will leave off the initial

values that I was talking about just a second go for right now.

So we went to any initial values just yet let's just get this component to show up on the screen and

then we'll come back and visit this initial values thing.

But this is where we will specify our initial values.

Last thing we need to do here down at the bottom we still want this map state to prop's so that we can

get some initial values for the form.

We still want to have the connect function here we still want to be able to fetch a stream.

The only other change that we need is to make sure that we pass in our action creator at it stream again.

That is the action creator that's going to make her request over to our API and update some given stream.

So I can make sure that on the Connect function right here I also put in at it stream.

OK so let's save this and we'll flip back with the browser and just make sure that we can get some form

appearing on the screen.

So when I flip back over I'm going to go back to my list of streams page and then click on edit for

one of these two streams right here doesn't matter which one sultry editing this one.

All right.

So I see my form right here it says edit a stream but as we said we don't have any initial values for

the form.