345.2. Setting Initial Values(Section 24: REST-Based React Apps)
So now we're going to come back to that topic of initial values.

So again we have stream edit it is showing a redux form wrapped version of stream form if we pass a

proper named initial values from stream edit down to stream form that initial values Propp will be used

as the initial values for the form.

So I'm going to open up stream at it same file we just had open.

When we create stream form right here I'm going to pass in a proper named initial values.

And again this is a very special property name with redux form.

I'm going to get myself a little bit of space here like so and now for initial values I'm going to put

in one set of curly braces and then remember we want to put in a second object here so I'm going to

put in a second set of curly braces.

Now the outside curly braces indicates that we are going to write some javascript expression inside

of our GSX.

The second set is indicating that we are creating a normal object.

And now I'm going to pass in a title we're going to eventually replace this with the title of the actual

string that we fetch.

But for right now I just want to give you a quick example of how this works.

So going to put in a title of edit me and a description of change me to about that

and I'll zoom out for a second so you can see that whole line.

Now because we are passing initial values down to a component that is wrapped by redux form these will

be used as the initial values inside of the form itself.

Now one thing that is going to hopefully connect the dots here just a little bit.

Why are we putting in an object with title and description.

Well we're using specifically those two property names because if you open up your stream form component

your stream form right here if you find the render method you will recall that we had two fields inside

of here.

The first field has a name of title.

The second field has a name of description.

So essentially when we show these field components right here they're going to look at any initial values

that were passed in from the parent component to this one.

They're going to see if that initial values property has a name of type or the property of title or

a property of description.

And if they do then those initial values will be used as the initial values for the field.

So that's the real connection between these two things.

OK.

So let's save all this and we'll flip back over to our application and we'll see that we do in fact

have these initial values here now.

So if I change the text here from at it me to edit this or whatever you want to change it to.

You'll see that we now have edit this there instead.