345.3. Setting Initial Values(Section 24: REST-Based React Apps)
And so in a very similar fashion we could remove this object right here.

And instead pass in our stream that we have loaded into this component so I can pass in this prop's

not stream.

Remember our stream is an object with a title and a description property.

So now the title and description off of our stream object will be used as the initial values for the

form.

Let's save this we can flip back over.

And now we'll see the initial value coming from the title the stream an initial value coming from the

description of the stream as well.

So just one more time really quickly here we are passing an initial valleys Propp down to stream form.

Remember stream form is technically rendered or wrapped by the redux form helper so that redux form

helper sees this prop of initial values it sees that it is an object with a title and a description

property.

And so when the stream form is rendered we have a field with a title and a field with a description.

And so those initial values are used as the initial values for a field number one and field number two

right here.

So now we should be able to make changes to this.

And when we submit the form it's going to call the callback handler that we had defined inside a stream

at the summit helper function right here.

Remember we defined on submit and we passed that down into our stream form right there.

And right now our handler is just doing that on a log.

So after making those changes I'm going to click on submit and we'll see our console log right here.

Now notice that the description and the tile have been updated properly.

But there's one little issue with the console log that we see right here.

One not so good little issue.

So let's take a quick pause right here.

When we come back the next section will talk about what is kind of off about this console lock.