346.2. Avoiding Changes to Properties(Section 24: REST-Based React Apps)
Well I'm going to show you another way of handling this that's going to be just a little bit more concise.

So I'm going to delete everything in between those curly braces and then at the very top of this file

We're going to import load ash.

Remember that is that helper library that we've been using quite a bit.

So we're going to use a function from Lotus to basically just pick out the values or the properties

from that stream object that we care about and that we actually want to pass down as initial values

to our form.

So back down at initial values right here I'm going to say underscore dot pick

and then I'm going to put in my stream object.

This dot props start Stream.

And then I will put in a list of all the different properties that I want to essentially pull out of

that object and pass down to initial values.

So in our case we want to pick out title and description.

And let me say this you can see the entire line very easily OK.

Let's test this out very quickly and we'll take a look at a quick example of how pick works just in

a little bit more greater detail.

So I going to save this I'm going to flip back over.

Now I'm back at my form right here and if I look at my redux dev tools and look at my state I can open

up loops form again.

I can look at my stream form I can look at values and I'll see that I only have the title and description

loaded up into redux form.

So now if I change these and submit the form again I see that my essentially kind of like change object

right here only has the title and description no longer is my idea inside there no longer is the user

ID inside there.

It's only the properties that I actually changed inside the form.

All right.

So again we did that by making use of this pick function from loadout.

So just really quickly I want to show you a quick example of what pick is really doing for us even though

I mean you can probably imagine what it's doing.

Let me give you a quick example.

Nonetheless someone to do my favorite thing hearing and a pull up in that little code editor again there

is the link to it if you want to follow along so inside if you're all put together some object like

say profile and I'll give it a name of Sam and age of 18.

And how about a favorite color.

Of green.

So now I can call underscore pic on profile and as the additional arguments into this thing I can list

out the different properties that I want to pull out of this profile object so I can say name like so

and you'll see that on a top right hand side I now get a new object returned.

That has only the name property out of profile.

This does not actually modify the profile object if I printout profile as well.

You'll see that profile itself is completely unchanged.

So profile creates a new object with de-listed properties.

So I could add on a third argument here of age and now I get name of Sam an age of 18.

So again we're doing the same exact thing inside of our form for our initial values right here we are

just picking out the form values that we care about and that we want to actually edit inside that form.

All right so I think that we probably get a good idea of what's going on here with the initial values.

Now let's take a quick pause.

When we come back the next section we're going to wrap up all this stream at stuff by making sure that

our Samit callback right here comes the appropriate action at Kreator to actually edit and update our

stream.

So quick pause and I'll see you in just a minute.