350.1. More on Using Portals(Section 25: Using React Portals)
In the last video we started putting together an HMO document for this little model example.

So let's keep working on it.

Right now I'm gonna flip back over to my HMO document and down inside of the body I'm going to place

a new div above the sidebar please make sure you make it above the sidebar and I'm going to give it

a class of motile as you might guess this is going to serve as the modal element that should be obscuring

the sidebar and the content inside of this div.

I'm going to place a div with a class of modal Dasch body and inside of that inner inner div I'll give

it an H one with some text like I am a modal like so.

Now why are we using two divs here.

Well the outer div right here is meant to serve as the kind of gray background that's going to obscure

all the content behind it the inner div with modal Dasch body is meant to represent this element right

here.

So the actual kind of white window of sorts that's going to show some content inside of it.

So usually we put these models together with the two separate elements one for the background one for

the foreground so to speak.

Now if we save that HMO document now and then refresh our page no we don't really see a whole lot going

on.

You'll notice that we now see I am some content shifted down.

The reason we don't see our content that says I'm a modal is that it's been obscured behind the sidebar

right now.

So we really need to add in some more CSSA before this is going to make any sense.

Let's look back over.