350.2. More on Using Portals(Section 25: Using React Portals)
All right inside of my style tag I'm going to add on some selectors for modal and modal dash body.

I'm going to start off first with modal so going to say modal.

I'm going to give this a height of 100 percent or with of 100 percent.

I'll give it a position of fixed a background color of gray and then a left of zero a top of zero.

Now I think that should be just about it for the background and we'll add in just a little bit for the

modal body as well.

I'm going to put the modal body selector above the modal just because it's going to make looking at

some the CSSA in just a moment a little bit easier for us.

So I'll say up here.

Motile Dasch body will have a background color of white.

I'll give it a margin.

Auto which is going to center this window on the screen horizontally that is not vertically.

Well give it a height of 30 percent and a width of 30 percent.

Now some of these rules in the long run are not going to work out super well for motile.

I'm just trying to get a example here on the window or inside of a browser.

Very quickly so I'm going to save this again.

Flip back over.

All right.

Now that looks a little bit better.

There's just one little issue.

I see the Modahl on screen.

So here's the kind of white content center window.

I see the gray background but you'll notice the issue here is that we've got the green.

I am a sidebar still on the left hand side.

So ideally we would get the side bar behind the motile and I shouldn't even really say idealy that's

just plain how it should work.

The sidebar should not be displayed on top of the motile.

So we need to figure out what is going on here and how we can fix this issue well to fix this issue.